perf: Move data download to client side
...to avoid sending the data back and forth.
